+++ integration_tests/test_terraform/tf_state.cfn/tfstate.yaml
@@ -8,6 +8,6 @@ stacks:
     template_path: templates/tf_state.yaml
 
 pre_destroy:
-  - path: runway.hooks.cleanup_s3.purge_bucket
+  - path: runway.cfngin.hooks.cleanup_s3.purge_bucket
     args:
       bucket_name: ${rxref ${namespace}-tf-state::TerraformStateBucketName}
