+++ pyproject.toml
@@ -40,7 +40,7 @@ concurrency = [
 ]
 omit = [
     "*/runway/aws_sso_botocore/*",  # TODO remove native support is added to botocore
-    "*/runway/hooks/staticsite/auth_at_edge/templates/*",
+    "*/runway/cfngin/hooks/staticsite/auth_at_edge/templates/*",
     "*/compat.py",
     "*/type_defs.py",
 ]
