+++ runway/blueprints/staticsite/dependencies.py
@@ -7,7 +7,7 @@ from awacs.aws import Allow, AWSPrincipal, Policy, Statement
 from troposphere import AccountId, Join, Output, cognito, s3
 
 from ...cfngin.blueprints.base import Blueprint
-from ...hooks.staticsite.auth_at_edge.client_updater import get_redirect_uris
+from ...cfngin.hooks.staticsite.auth_at_edge.client_updater import get_redirect_uris
 from ...module.staticsite.utils import add_url_scheme
 
 LOGGER = logging.getLogger(__name__)
