+++ runway/cfngin/hooks/cleanup_s3.py
@@ -1,5 +1,4 @@
 """CFNgin hook for cleaning up resources prior to CFN stack deletion."""
-# TODO move to runway.cfngin.hooks on next major release
 from __future__ import annotations
 
 import logging
@@ -8,7 +7,7 @@ from typing import TYPE_CHECKING, Any
 from botocore.exceptions import ClientError
 
 if TYPE_CHECKING:
-    from ..context import CfnginContext
+    from ...context import CfnginContext
 
 LOGGER = logging.getLogger(__name__)
