+++ runway/cfngin/hooks/staticsite/__init__.py
