+++ runway/cfngin/hooks/staticsite/auth_at_edge/__init__.py
