+++ runway/cfngin/hooks/staticsite/auth_at_edge/callback_url_retriever.py
@@ -10,7 +10,7 @@ import logging
 from typing import TYPE_CHECKING, Any, Dict, Optional
 
 if TYPE_CHECKING:
-    from ....context import CfnginContext
+    from .....context import CfnginContext
 
 LOGGER = logging.getLogger(__name__)
