+++ runway/cfngin/hooks/staticsite/auth_at_edge/client_updater.py
@@ -11,7 +11,7 @@ import logging
 from typing import TYPE_CHECKING, Any, Dict, List, Optional
 
 if TYPE_CHECKING:
-    from ....context import CfnginContext
+    from .....context import CfnginContext
 
 LOGGER = logging.getLogger(__name__)
