+++ runway/cfngin/hooks/staticsite/auth_at_edge/templates/__init__.py
