+++ runway/cfngin/hooks/staticsite/auth_at_edge/templates/check_auth/__init__.py
