+++ runway/cfngin/hooks/staticsite/auth_at_edge/templates/http_headers/__init__.py
