+++ runway/cfngin/hooks/staticsite/auth_at_edge/templates/parse_auth/__init__.py
