+++ runway/cfngin/hooks/staticsite/auth_at_edge/templates/refresh_auth/__init__.py
