+++ runway/cfngin/hooks/staticsite/auth_at_edge/templates/shared_jose.py
