+++ runway/cfngin/hooks/staticsite/auth_at_edge/templates/sign_out/__init__.py
