+++ runway/cfngin/hooks/staticsite/auth_at_edge/user_pool_id_retriever.py
