+++ tests/unit/cfngin/hooks/staticsite/__init__.py
