--- tests/unit/hooks/staticsite/__init__.py
@@ -1,1 +1,0 @@
-"""Empty init for python import traversal."""
